add timestamp check

+++ nursery_pdftocsv.py
@@ -4,6 +4,8 @@ import csv
 import os.path
 
 def proc(input, output):
+	if os.path.exists(output) and os.stat(output).st_mtime > os.stat(input).st_mtime:
+		return
 	rs = pte.table_to_list(pte.process_page(input, "1"), 1)
 	w = csv.writer(open(output, "w"))
 	w.writerows(rs[1])
